Stop a KPI reconfiguration from carrying stale totals, and unbrick the endTime read
Two findings from the `1875ba6..ac7eb44` review.

`setKpiConfig` rewrote the whole `KpiConfig` but left `lastScannedBlock` and every
`verifiedTotals` entry untouched. Correcting a KPI configured against the wrong contract
therefore resumed the relayer at `N+1`, so `[windowStartBlock, N]` was never rescanned for
the right event — while the totals folded from the wrong one stayed live as the ceiling a
claim is trimmed to. Raising `windowStartBlock` left the same gap in reverse, and changing
`scale` re-denominated every stored total retroactively.

`KpiConfig` now carries an `epoch` that is part of every total's storage key, bumped
whenever the watched event changes and left alone when only `windowEndBlock` moves — the
one case replacement exists for. A bump abandons every figure observed under the old config
in a single write and resets the checkpoint. Clearing the entries instead is not available:
`verifiedTotals` is a mapping with no enumerable key set, so a user who only ever appeared
under the wrong config would never be overwritten at all.

`AttributionRegistry._requireCampaignOpen` decoded `endTime()` as `uint64`, which reverts on
a non-zero upper word. The `status()` read four lines below was deliberately decoded as
`uint256` for exactly this reason and says so; the reasoning had been applied to one read and
not its neighbour. A registrant returning a full 32-byte value could therefore brick every
`storeTouch` naming it, contradicting the documented reading that a campaign answering
neither call is simply unbounded. Both that read and `attributionWindow` in
`_effectiveMaxDuration` now decode as `uint256` and compare numerically, and `CampaignOver`
carries the full word so it cannot report a garbage endTime as a plausible one.

414 forge tests pass. The verifier's contract-level and struct docs also carry an in-flight
comment trim from a concurrent session in this working tree.

+++ src/attribution/AttributionRegistry.sol
@@ -152,12 +152,16 @@ contract AttributionRegistry is IAttributionRegistry, EIP712 {
         // A non-campaign returns no data; a conforming campaign returns exactly one word.
         if (!ok || data.length != 32) return maxTouchDuration;
 
-        uint64 window = abi.decode(data, (uint64));
+        // Decoded as uint256, not uint64: `abi.decode` into the narrower type reverts on a dirty
+        // upper word, which would let a registrant returning a full 32-byte value brick every touch
+        // naming it. Comparing numerically instead means an oversized answer simply clamps to the cap,
+        // which is the documented reading — see `_requireCampaignOpen` for the same decision.
+        uint256 window = abi.decode(data, (uint256));
         // Campaign.sol rejects a zero window at construction, so zero here means "not a campaign"
         // rather than "no attribution allowed" — falling through to the cap keeps a decoding
         // surprise from bricking every touch for that address.
         if (window == 0 || window > maxTouchDuration) return maxTouchDuration;
-        return window;
+        return uint64(window);
     }
 
     /// @dev Reverts unless `campaign` can still accrue creditable work.
@@ -184,10 +188,16 @@ contract AttributionRegistry is IAttributionRegistry, EIP712 {
     function _requireCampaignOpen(address campaign, uint64 nowTs) private view {
         (bool okEnd, bytes memory endData) = campaign.staticcall(abi.encodeCall(ICampaignWindow.endTime, ()));
         if (okEnd && endData.length == 32) {
-            uint64 end = abi.decode(endData, (uint64));
+            // Decoded as uint256, not as the declared uint64, for the same reason `status` below is
+            // not decoded as its enum: `abi.decode` into the narrower type reverts on a non-zero
+            // upper word, so a registrant answering with a full 32-byte value could brick every touch
+            // naming it — the opposite of the "answers neither call is simply unbounded" reading
+            // above. Comparing numerically keeps an oversized answer merely far in the future, and
+            // the error carries the full word so it cannot report a garbage endTime as a plausible one.
+            uint256 end = abi.decode(endData, (uint256));
             // Campaign.sol rejects `endTime <= block.timestamp` at construction, so zero is "not a
             // campaign" rather than "already over" — the same reading as a zero window above.
-            if (end != 0 && nowTs > end) revert CampaignOver(end, nowTs);
+            if (end != 0 && uint256(nowTs) > end) revert CampaignOver(end, nowTs);
         }
 
         (bool okStatus, bytes memory statusData) =
